+++ includes/modules/payment/emspay_applepay.php
@@ -123,7 +123,7 @@ class emspay_applepay extends emspayGateway
         global $order, $messageStack;
 
         try {
-            $emsOrder = $this->emspay->createOrder([
+            $emsOrder = $this->emspay->createOrder(array_filter([
                 'amount' => $this->gerOrderTotalInCents($order),       // amount in cents
                 'currency' => $this->getCurrency($order),              // currency
                 'description' => $this->getOrderDescription(),         // order description
@@ -137,7 +137,7 @@ class emspay_applepay extends emspayGateway
                         'payment_method' => 'apple-pay',
                     ]
                 ]
-            ]);
+            ]));
 
             if ($emsOrder['status'] == 'error') {
                 $messageStack->add_session('checkout_payment', MODULE_PAYMENT_EMSPAY_APPLEPAY_ERROR_TRANSACTION, 'error');